requirements.txt
streamlit
ollama
duckduckgo-search
httpx
python-dotenv
Pillow
PyPDF2
sentence-transformers
chromadb
